requirements.txt
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pytest==8.3.3
typer==0.12.5